tests/wpt/tests/html/semantics/forms/customizable-combobox/customizable-combobox-filtered-pseudo.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://open-ui.org/components/combobox.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<input list=datalist style="appearance:base">
<datalist id=datalist style="appearance:base">
  <option class=one>one</option>
  <option class=two>two</option>
  <option class=three>three</option>
</datalist>

<script>
const input = document.querySelector('input');
const optionOne = document.querySelector('.one');
const optionTwo = document.querySelector('.two');
const optionThree = document.querySelector('.three');

function checkOptions(includedOptions, message) {
  for (const option of document.querySelectorAll('option')) {
    if (includedOptions.includes(option)) {
      assert_false(
        option.matches(':filtered'),
        `${message}: option ${option.className} should not match :filtered.`);
      assert_not_equals(
        getComputedStyle(option).display,
        'none',
        `${message}: option ${option.className} should not be display:none.`);
    } else {
      assert_true(
        option.matches(':filtered'),
        `${message}: option ${option.className} should match :filtered.`);
      assert_equals(
        getComputedStyle(option).display,
        'none',
        `${message}: option ${option.className} should be display:none.`);
    }
  }
}

function pressKey(keyCode) {
  return (new test_driver.Actions()
    .keyDown(keyCode)
    .keyUp(keyCode))
    .send();
}

promise_test(async () => {
  checkOptions([optionOne, optionTwo, optionThree],
    'No options should be filtered out at the start of the test');

  input.focus();
  await pressKey('t');
  checkOptions([optionTwo, optionThree],
    'First option should be filtered out after typing "t"');

  await pressKey('w');
  checkOptions([optionTwo],
    'Third option should be filtered out after typing "tw"');

  await pressKey('z');
  checkOptions([],
    'All options should be filtered out after typing "twz"');

  const backspace = '\uE003';
  await pressKey(backspace);
  checkOptions([optionTwo],
    'Second option should become unfiltered after pressing backspace.');

  input.value = '';
  checkOptions([optionOne, optionTwo, optionThree],
    'No options should be filtered out after setting input.value to the empty string.');

  input.value = 'tw';
  checkOptions([optionTwo],
    'Only option two should match when setting input.value to "tw".');
}, 'Customizable combobox option:filtered behavior');
</script>
